+++ database/migrations/2021_07_27_034747_create_fasilitas_umum.php
@@ -14,7 +14,7 @@ class CreateFasilitas extends Migration
     public function up()
     {
         Schema::create('fasilitas', function (Blueprint $table) {
-            $table->bigIncrements('id_fasilitas_umum');
+            $table->Integer('id_fasilitas_umum');
             $table->string('nama_fasilitas');
             $table->string('kategori_fasilitas');
             $table->Integer('jumlah_fasilitas');
