+++ database/migrations/2021_07_27_065327_create_kelurahan_table.php
@@ -14,7 +14,7 @@ class CreateKelurahanTable extends Migration
     public function up()
     {
         Schema::create('kelurahan', function (Blueprint $table) {
-            $table->bigIncrements('id_kelurahan');
+            $table->Integer('id_kelurahan');
             $table->string('nama_kelurahan');
             $table->Integer('id_kecamatan')->nullable();
             $table->Integer('id_kota')->nullable();
