+++ database/migrations/2021_07_27_072904_create_kecamatan_table.php
@@ -14,7 +14,7 @@ class CreateKecamatanTable extends Migration
     public function up()
     {
         Schema::create('kecamatan', function (Blueprint $table) {
-            $table->bigIncrements('id_kecamatan');
+            $table->Integer('id_kecamatan');
             $table->string('nama_kecamatan');
             $table->Integer('id_kota');
             $table->timestamps();
