+++ database/migrations/2021_09_07_004029_create_detail_transaksi_pembayaran.php
@@ -14,7 +14,7 @@ class CreateDetailTransaksiPembayaran extends Migration
     public function up()
     {
         Schema::create('detail_transaksi_pembayaran', function (Blueprint $table) {
-            $table->bigIncrements('id_detail_transaksi_pembayaran');           
+            $table->Integer('id_detail_transaksi_pembayaran');           
             $table->date('bulan');
             $table->Integer('harga');
             $table->Integer('id_transaksi_pembayaran');
