+++ database/migrations/2021_09_19_084932_create_syarat_mendaftar_table.php
@@ -14,7 +14,7 @@ class CreateSyaratMendaftarTable extends Migration
     public function up()
     {
         Schema::create('syarat_mendaftar', function (Blueprint $table) {
-            $table->bigIncrements('id_syarat_mendaftar');
+            $table->Integer('id_syarat_mendaftar');
             $table->string('syarat_mendaftar');
             $table->timestamps();
         });
